Major Progress; Shall Be Continued
Really late on this commit as it should have been done yesterday, but I finally got a lot of work done on the index page that shall be merged with the index. The code itself needs a bit more polishing before it can rely on its own (that's what I'm about to do), but it does come as a fully responsive product that shouldn't take much longer if I really pull at it.