Minor Modifications Added; Index Complete
The header list has been successfully implemented to its full capacity and  as such, leads the about part of the main page the only part left. I don't assume it to take long at all when it's really just adding on, so I'll get right to work on that tomorrow and hopefully finish it as well.